+++ admin/class-tmsm-aquos-spa-booking-admin.php
@@ -510,7 +510,7 @@ class Tmsm_Aquos_Spa_Booking_Admin {
 		if(is_multisite() && get_current_blog_id() === 6) { // Rennes
 			switch ( $aquos_id ) {
 				case 338: // Parcours Aquatonic
-					$price = 20;
+					$price = 21;
 					break;
 				case 191: // Modelage nuque et cuir chevelu
 					$price = 18;